requirements.txt
python-dotenv==1.0.1
pandas==2.2.2
spacy==3.7.4
scikit-learn==1.5.1
numpy==1.26.4
langchain==0.2.9
crewai==0.51.1
tiktoken==0.7.0
